refactor(settings): extract SMTP form into SettingsEmail.svelte
